Refactor : mise en place d'un module de membres
