+++ src/app/membres/containers/membres/members-container.component.html
@@ -9,4 +9,3 @@
     <h2>Revenez plus tard ;)</h2>
   </div>
 </section>
-<app-footer></app-footer>
