+++ src/app/membres/containers/membres/members-container.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import {SeoService} from "../../common/services/seo/seo.service";
+import {SeoService} from "../../../common/services/seo/seo.service";
 
 @Component({
   selector: 'app-membres',
-  templateUrl: './membres.component.html',
-  styleUrls: ['./membres.component.scss']
+  templateUrl: './members-container.component.html',
+  styleUrls: ['./members-container.component.scss']
 })
-export class MembresComponent implements OnInit {
+export class MembersContainer implements OnInit {
 
   constructor(private seo:SeoService) {
     this.seo.generateTags({
