+++ src/app/nous-rejoindre/containers/nous-rejoindre/nous-rejoindre.component.html
@@ -35,7 +35,7 @@
               règles</strong>. En cas de non respect de celles-ci, les joueurs concernés seront sanctionnés sur-le-champ.
             </p>
             <div class="call-to-action-container">
-              <a mat-stroked-button color="primary" href="/reglement" routerLink="/reglement">
+              <a mat-stroked-button color="primary" href="/nous-rejoindre/reglement" routerLink="/nous-rejoindre/reglement">
                 Réglement
               </a>
             </div>
