+++ src/app/nous-rejoindre/containers/reglement/reglement.component.html
@@ -55,4 +55,3 @@
 
   </div>
 </section>
-<app-footer></app-footer>
