+++ src/app/nous-rejoindre/nous-rejoindre-routing.module.ts
@@ -8,6 +8,7 @@ import {
   NousRejoindreFormCounselComponent
 } from "./containers/nous-rejoindre-form-counsel/nous-rejoindre-form-counsel.component";
 import {NousRejoindreFormComponent} from "./containers/nous-rejoindre-form/nous-rejoindre-form.component";
+import {ReglementComponent} from "./containers/reglement/reglement.component";
 
 
 
@@ -16,6 +17,7 @@ const routes : Routes = [
   { path:'1', component:NousRejoindreFormAdhesionComponent, data: { animation: 'nous-rejoindre-1' }},
   { path:'2', component:NousRejoindreFormCounselComponent, data: { animation: 'nous-rejoindre-2' }},
   { path:'3', component:NousRejoindreFormComponent, data: { animation: 'nous-rejoindre-3' }},
+  {title:'Litopia - Réglement', path:'reglement', component:ReglementComponent, data: { animation: 'reglement' }},
 ]
 @NgModule({
   imports: [RouterModule.forChild(routes)],
